+++ CHANGELOG.md
@@ -3,13 +3,19 @@
 This log documents the significant changes for each release.
 This project follows [Semantic Versioning](http://semver.org/).
 
-## [3.17.2] - 2021-09-09
+## [3.18.1] - 2021-09-10
 ### Fixed
 - Adding a new resource tab in the Pull data for cohort step after or when
   loading Observations data threw exceptions that blocked further operation
   of the application.
 - Issue with displaying an empty table when changing the active tab in
   the Pull data for cohort step.
+
+## [3.18.0] - 2021-09-09
+### Added
+- Using XLSX file(s) to configure search parameters and resource table columns.
+### Fixed
+- Switching to the View Cohort step after loading criteria and cohort.
 
 ## [3.17.1] - 2021-09-08
 ### Changed